+++ src/client/scripts/esm/game/chess/selection.js
@@ -24,6 +24,8 @@ import draganimation from '../rendering/draganimation.js';
 import space from '../misc/space.js';
 import preferences from '../../components/header/preferences.js';
 import gameslot from './gameslot.js';
+import gameloader from './gameloader.js';
+import movesendreceive from '../misc/onlinegame/movesendreceive.js';
 // Import End
 
 /**
@@ -359,7 +361,7 @@ function moveGamefilePiece(coords) {
 	const move = movesequence.makeMove(gameslot.getGamefile(), moveDraft);
 	// Don't animate the main piece if it's being dragged, but still animate secondary pieces affected by the move (like the rook in castling).
 	movesequence.animateMove(move, true, animateMain);
-	onlinegame.sendMove();
+	movesendreceive.sendMove();
 
 	unselectPiece();
 }
